Update: Chat - make push message function receive a simple object and then create a chat message class instance from it, add type class suffix property to chat message, add chat message system modifier class, cleanup code
The change to chat push message function is to make it easier to use it externally without the need to import the chat message class. It is now done internally.
Messages now have the "system" type class suffix to differentiate it from regular messages. It can be used for system messages like when a user leaves the chat.